updating staging migrations to not crash if there are some tables present already

+++ db/migrate_staging/20110908094408_add_street_zip_and_city_to_sta_executors.rb
@@ -1,10 +1,10 @@
 class AddStreetZipAndCityToStaExecutors < ActiveRecord::Migration
   def self.up
-    add_column :sta_executors, :street, :string
-    add_column :sta_executors, :zip, :string
-    add_column :sta_executors, :city, :string
+    add_column :sta_executors, :street, :string unless column_exists? :sta_executors, :street
+    add_column :sta_executors, :zip, :string unless column_exists? :sta_executors, :zip
+    add_column :sta_executors, :city, :string unless column_exists? :sta_executors, :city
     
-    remove_column :sta_executors, :address
+    remove_column :sta_executors, :address if column_exists? :sta_executors, :address
   end
 
   def self.down
